+++ homecare/src/main/java/jaega/homecare/domain/workLog/dto/res/GetCaregiverMatchesByMonth.java
@@ -1,0 +1,22 @@
+package jaega.homecare.domain.workLog.dto.res;
+
+import jaega.homecare.domain.workLog.entity.WorkStatus;
+import jaega.homecare.domain.users.entity.ServiceType;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.Set;
+import java.util.UUID;
+
+public record GetCaregiverMatchesByMonth(
+        UUID workLogId,
+        UUID caregiverId,
+        String caregiverName,
+        LocalDate workDate,
+        LocalTime startTime,
+        LocalTime endTime,
+        Set<ServiceType> serviceType,
+        String address,
+        WorkStatus status
+) {
+}
