+++ homecare/src/main/java/jaega/homecare/domain/workLog/dto/res/GetMonthlyPaymentResponse.java
@@ -1,0 +1,10 @@
+package jaega.homecare.domain.workLog.dto.res;
+
+import java.math.BigDecimal;
+
+public record GetMonthlyPaymentResponse(
+        int year,
+        int month,
+        BigDecimal totalAmount
+) {
+}
