+++ homecare/src/main/java/jaega/homecare/domain/workLog/dto/res/WorkPlaceDistribution.java
@@ -1,0 +1,9 @@
+package jaega.homecare.domain.workLog.dto.res;
+
+import jaega.homecare.domain.users.entity.ServiceType;
+
+public record WorkPlaceDistribution (
+        ServiceType serviceType,
+        Long count,
+        Double percent
+) {}
