+++ homecare/src/main/java/jaega/homecare/domain/workLog/entity/WorkStatus.java
@@ -1,0 +1,7 @@
+package jaega.homecare.domain.workLog.entity;
+
+public enum WorkStatus {
+    PLANNED,       // 예정
+    COMPLETED,     // 완료
+    CANCELLED      // 취소됨
+}
