--- homecare/src/main/java/jaega/homecare/domain/workMatch/controller/WorkMatchControllerImpl.java
@@ -1,90 +1,0 @@
-package jaega.homecare.domain.workMatch.controller;
-
-import jaega.homecare.domain.workMatch.dto.res.*;
-import jaega.homecare.domain.workMatch.entity.WorkStatus;
-import jaega.homecare.domain.workMatch.service.query.WorkMatchQueryService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.time.LocalDate;
-import java.util.List;
-import java.util.UUID;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/api/workMatch")
-public class WorkMatchControllerImpl implements WorkMatchController{
-    private final WorkMatchQueryService workMatchQueryService;
-
-
-    // 정산 페이지
-
-    /**
-     * 센터에 해당하는 요양보호사 근무 현황 조회
-     */
-    @Override
-    public ResponseEntity<List<GetCaregiverWorkResponse>> getCaregiverWorkList(
-            @PathVariable UUID centerId,
-            @RequestParam(required = false) WorkStatus status,
-            @RequestParam(required = false) Integer year,
-            @RequestParam(required = false) Integer month
-    ) {
-        List<GetCaregiverWorkResponse> result = workMatchQueryService.getCaregiverWorkList(centerId, status, year, month);
-        return ResponseEntity.ok(result);
-    }
-
-    @Override
-    public ResponseEntity<List<GetCaregiverWorkResponse>> getCaregiverWorkListByCaregiver(
-            @PathVariable UUID caregiverId,
-            @RequestParam(required = false) WorkStatus status,
-            @RequestParam(required = false) Integer year,
-            @RequestParam(required = false) Integer month
-    ) {
-        List<GetCaregiverWorkResponse> result = workMatchQueryService.getCaregiverWorkListByCaregiver(caregiverId, status, year, month);
-        return ResponseEntity.ok(result);
-    }
-
-    @Override
-    public List<GetMonthlyPaymentResponse> getMonthlyPaid(@RequestParam UUID centerId) {
-        return workMatchQueryService.getMonthlyPaidSettlements(centerId);
-    }
-
-    @Override
-    public List<GetDailyUnsettledResponse> getDailyUnsettled(@RequestParam UUID centerId) {
-        return workMatchQueryService.getDailyUnsettledCount(centerId);
-    }
-
-    @Override
-    public GetSettlementCenterSummaryResponse getSettlementSummary(@RequestParam UUID centerId) {
-        return workMatchQueryService.getSettlementSummary(centerId);
-    }
-
-    @Override
-    public ResponseEntity<GetCaregiverSettlementSummaryResponse> getCaregiverSettlementSummary(
-            @PathVariable UUID caregiverId
-    ) {
-        return ResponseEntity.ok(workMatchQueryService.getCaregiverSettlementSummary(caregiverId));
-    }
-
-    @Override
-    public ResponseEntity<GetWorkMatchResponse> getWorkMatch(@PathVariable UUID workMatchId) {
-        GetWorkMatchResponse responses = workMatchQueryService.findWorkMatch(workMatchId);
-        return ResponseEntity.ok(responses);
-    }
-
-    @Override
-    public ResponseEntity<List<GetWorkMatchByDateResponse>> getWorkMatchByWorkDay(@RequestParam UUID centerId, @RequestParam LocalDate workDate) {
-        List<GetWorkMatchByDateResponse> response = workMatchQueryService.getWorkMatchByDate(centerId, workDate);
-        return ResponseEntity.ok(response);
-    }
-
-    @Override
-    public ResponseEntity<List<GetWorkMatchByPaid>> getWorkMatchByPaid(@RequestParam UUID centerId, Boolean isPaid) {
-        List<GetWorkMatchByPaid> response = workMatchQueryService.getWorkMatchByPaid(centerId, isPaid);
-        return ResponseEntity.ok(response);
-    }
-}
