--- homecare/src/main/java/jaega/homecare/domain/workMatch/dto/req/CreateWorkLogRequest.java
@@ -1,13 +1,0 @@
-package jaega.homecare.domain.workMatch.dto.req;
-
-import jaega.homecare.domain.workMatch.entity.WorkMatch;
-
-import java.time.LocalTime;
-
-public record CreateWorkLogRequest(
-        WorkMatch workMatch,
-        LocalTime workTime_start,
-        LocalTime workTime_end,
-        Double distanceLog
-) {
-}
