--- homecare/src/main/java/jaega/homecare/domain/workMatch/dto/req/CreateWorkMatchRequest.java
@@ -1,16 +1,0 @@
-package jaega.homecare.domain.workMatch.dto.req;
-
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.UUID;
-
-public record CreateWorkMatchRequest(
-        UUID caregiverId,
-        LocalTime workStartTime,
-        LocalTime workEndTime,
-        LocalDate workDate,
-        String workAddress,
-        Double distanceLog
-) {
-}
