--- homecare/src/main/java/jaega/homecare/domain/workMatch/dto/req/GetWorkLogRequest.java
@@ -1,8 +1,0 @@
-package jaega.homecare.domain.workMatch.dto.req;
-
-import java.util.UUID;
-
-public record GetWorkLogRequest (
-        UUID workMatchId
-) {
-}
