--- homecare/src/main/java/jaega/homecare/domain/workMatch/dto/res/GetDashboardSettlementResponse.java
@@ -1,10 +1,0 @@
-package jaega.homecare.domain.workMatch.dto.res;
-
-import java.math.BigDecimal;
-
-public record GetDashboardSettlementResponse(
-        BigDecimal totalSettledAmount,
-        Long unsettledCount,
-        Long fraudAlertCount  // 부정행위 알림 건수
-) {
-}
