--- homecare/src/main/java/jaega/homecare/domain/workMatch/dto/res/GetMonthlyPaymentResponse.java
@@ -1,10 +1,0 @@
-package jaega.homecare.domain.workMatch.dto.res;
-
-import java.math.BigDecimal;
-
-public record GetMonthlyPaymentResponse(
-        int year,
-        int month,
-        BigDecimal totalAmount
-) {
-}
