--- homecare/src/main/java/jaega/homecare/domain/workMatch/dto/res/GetWorkMatchByPaid.java
@@ -1,11 +1,0 @@
-package jaega.homecare.domain.workMatch.dto.res;
-
-import java.time.LocalDate;
-import java.util.UUID;
-
-public record GetWorkMatchByPaid(
-        UUID workMatchId,
-        LocalDate workingDate,
-        String caregiverName
-) {
-}
