--- homecare/src/main/java/jaega/homecare/domain/workMatch/dto/res/GetWorkMatchResponse.java
@@ -1,13 +1,0 @@
-package jaega.homecare.domain.workMatch.dto.res;
-
-import java.time.LocalTime;
-import java.util.UUID;
-
-public record GetWorkMatchResponse(
-        UUID workMatchId,
-        LocalTime workStartTime,
-        LocalTime workEndTime,
-        Double distanceLog,
-        Boolean isPaid
-) {
-}
