--- homecare/src/main/java/jaega/homecare/domain/workMatch/dto/res/WorkPlaceDistribution.java
@@ -1,9 +1,0 @@
-package jaega.homecare.domain.workMatch.dto.res;
-
-import jaega.homecare.domain.users.entity.ServiceType;
-
-public record WorkPlaceDistribution (
-        ServiceType serviceType,
-        Long count,
-        Double percent
-) {}
