--- homecare/src/main/java/jaega/homecare/domain/workMatch/entity/WorkStatus.java
@@ -1,7 +1,0 @@
-package jaega.homecare.domain.workMatch.entity;
-
-public enum WorkStatus {
-    PLANNED,       // 예정
-    COMPLETED,     // 완료
-    CANCELLED      // 취소됨
-}
