--- homecare/src/main/java/jaega/homecare/domain/workMatch/repository/WorkMatchQueryRepository.java
@@ -1,636 +1,0 @@
-package jaega.homecare.domain.workMatch.repository;
-
-import com.querydsl.core.BooleanBuilder;
-import com.querydsl.core.Tuple;
-import com.querydsl.core.types.Projections;
-import com.querydsl.core.types.dsl.Expressions;
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jaega.homecare.domain.workMatch.dto.res.*;
-import jaega.homecare.domain.workMatch.entity.QWorkMatch;
-import jaega.homecare.domain.workMatch.entity.WorkMatch;
-import jaega.homecare.domain.workMatch.entity.WorkStatus;
-import jaega.homecare.domain.caregiver.entity.Caregiver;
-import jaega.homecare.domain.caregiver.entity.QCaregiver;
-import jaega.homecare.domain.caregiver.repository.CaregiverRepository;
-import jaega.homecare.domain.caregiverCenter.entity.CaregiverStatus;
-import jaega.homecare.domain.caregiverCenter.entity.QCaregiverCenter;
-import jaega.homecare.domain.users.entity.QUser;
-import jaega.homecare.domain.users.entity.ServiceType;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.*;
-import java.util.stream.Collectors;
-
-@Repository
-@RequiredArgsConstructor
-public class WorkMatchQueryRepository {
-    private final JPAQueryFactory queryFactory;
-    private final CaregiverRepository caregiverRepository;
-
-    // 센터에 등록된 요양보호사 정산 내역 조회
-    public List<GetCaregiverMatchesByMonth> findWorkMatchByMonth(UUID centerId, int year, int month, Integer day) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QUser user = QUser.user;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        LocalDate startDate;
-        LocalDate endDate;
-
-        if (day != null) {
-            startDate = LocalDate.of(year, month, day);
-            endDate = startDate;
-        } else {
-            startDate = LocalDate.of(year, month, 1);
-            endDate = startDate.withDayOfMonth(startDate.lengthOfMonth());
-        }
-
-        // 1. 기본 정보 조회 (serviceTypes 제외)
-        List<GetCaregiverMatchesByMonth> baseList = queryFactory
-                .select(Projections.constructor(
-                        GetCaregiverMatchesByMonth.class,
-                        workMatch.workMatchId,
-                        caregiver.caregiverId,
-                        user.name,
-                        workMatch.workDate,
-                        workMatch.workStartTime,
-                        workMatch.workEndTime,
-                        Expressions.constant(Collections.emptySet()),
-                        caregiver.address,
-                        workMatch.status
-                ))
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiver.user, user)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(
-                        workMatch.workDate.between(startDate, endDate),
-                        caregiverCenter.center.centerId.eq(centerId)
-                )
-                .orderBy(workMatch.workDate.desc(), workMatch.createdAt.desc())
-                .fetch();
-
-        // 2. caregiverId 추출
-        Set<UUID> caregiverIds = baseList.stream()
-                .map(GetCaregiverMatchesByMonth::caregiverId)
-                .collect(Collectors.toSet());
-
-        if (caregiverIds.isEmpty()) {
-            return baseList;
-        }
-
-        // 3. serviceTypes 별도 조회
-        List<Object[]> rows = caregiverRepository.findServiceTypesByCaregiverIds(caregiverIds);
-
-        // 4. Map<Long, Set<ServiceType>> 변환
-        Map<UUID, Set<ServiceType>> serviceTypeMap = rows.stream()
-                .collect(Collectors.groupingBy(
-                        row -> (UUID) row[0],
-                        Collectors.mapping(row -> (ServiceType) row[1], Collectors.toSet())
-                ));
-
-        // 5. DTO 재생성
-        return baseList.stream()
-                .map(base -> new GetCaregiverMatchesByMonth(
-                        base.workMatchId(),
-                        base.caregiverId(),
-                        base.caregiverName(),
-                        base.workDate(),
-                        base.startTime(),
-                        base.endTime(),
-                        serviceTypeMap.getOrDefault(base.caregiverId(), Collections.emptySet()),
-                        base.address(),
-                        base.status()
-                ))
-                .toList();
-    }
-
-    // 매칭 알고리즘 사전 필터링
-    public List<WorkMatch> findOverlappingWorkMatch(
-            Caregiver caregiver,
-            LocalDate date,
-            LocalTime startTime,
-            LocalTime endTime
-    ) {
-        QWorkMatch wm = QWorkMatch.workMatch;
-
-        return queryFactory
-                .selectFrom(wm)
-                .where(
-                        wm.caregiver.eq(caregiver),
-                        wm.workDate.eq(date),
-                        wm.status.eq(WorkStatus.PLANNED),
-                        wm.workStartTime.lt(endTime),
-                        wm.workEndTime.gt(startTime)
-                )
-                .fetch();
-    }
-
-    /**
-     * 대시보드의 통계를 위한 api
-     */
-
-    // 오늘 근무하는 요양보호사 수 조회
-    public Long countCaregiversWorkingToday(UUID centerId) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        LocalDate today = LocalDate.now();
-
-        return queryFactory
-                .select(workMatch.caregiver.countDistinct())
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(
-                        caregiverCenter.center.centerId.eq(centerId)
-                                .and(workMatch.workDate.eq(today))
-                                .and(workMatch.status.in(WorkStatus.PLANNED, WorkStatus.COMPLETED))
-                )
-                .fetchOne();
-    }
-
-    // 오늚 미배정된 요양보호사 수 조회
-    public Long countUnassignedCaregiversToday(UUID centerId) {
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        LocalDate today = LocalDate.now();
-
-        // 오늘 배정된 caregiverId 서브쿼리
-        List<UUID> assignedCaregiverIds = queryFactory
-                .select(workMatch.caregiver.caregiverId)
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(workMatch.workDate.eq(today))
-                .fetch();
-
-        // 전체 센터 소속 caregiver 중 오늘 미배정인 수
-        return queryFactory
-                .select(caregiver.count())
-                .from(caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(
-                        caregiverCenter.center.centerId.eq(centerId)
-                                .and(caregiver.caregiverId.notIn(assignedCaregiverIds))
-                )
-                .fetchOne();
-    }
-
-    // 배정 대기인 요양보호사 수 조회
-    public Long countWaitingApplicants(UUID centerId) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        LocalDate today = LocalDate.now();
-
-        return queryFactory
-                .select(workMatch.count())
-                .from(workMatch)
-                .leftJoin(workMatch.caregiver, caregiver)
-                .leftJoin(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(
-                        workMatch.status.eq(WorkStatus.PLANNED)
-                                .and(workMatch.workDate.eq(today))
-                                .and(
-                                        workMatch.caregiver.isNull()
-                                                .or(caregiverCenter.center.centerId.eq(centerId))
-                                )
-                )
-                .fetchOne();
-    }
-
-    // 요양 보호사 대시보드의 근무지 별 분포 통계 조회
-    public List<WorkPlaceDistribution> getWorkPlaceDistributionByServiceType(UUID centerId) {
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        // center 소속 활성 보호사 + serviceTypes 조회
-        List<Tuple> rows = queryFactory
-                .select(caregiver, caregiver.serviceTypes)
-                .from(caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(
-                        caregiverCenter.center.centerId.eq(centerId),
-                        caregiverCenter.status.eq(CaregiverStatus.ACTIVE)
-                )
-                .fetch();
-
-        if (rows.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        // ServiceType별 카운트 계산
-        Map<ServiceType, Long> serviceTypeCount = new HashMap<>();
-        for (Tuple row : rows) {
-            Set<ServiceType> serviceTypes = row.get(caregiver.serviceTypes);
-            if (serviceTypes != null) {
-                for (ServiceType st : serviceTypes) {
-                    serviceTypeCount.put(st, serviceTypeCount.getOrDefault(st, 0L) + 1);
-                }
-            }
-        }
-
-        // 총합 계산
-        long total = serviceTypeCount.values().stream().mapToLong(Long::longValue).sum();
-
-        // DTO 변환
-        return serviceTypeCount.entrySet().stream()
-                .map(entry -> new WorkPlaceDistribution(
-                        entry.getKey(),
-                        entry.getValue(),
-                        total > 0 ? (double) entry.getValue() * 100 / total : 0.0
-                ))
-                .toList();
-    }
-
-    // 정산 페이지
-
-    // 센터에 등록된 요양보호사 정산 금액, 건수 조회
-    public GetSettlementCenterSummaryResponse getSettlementCenterSummary(UUID centerId) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        List<Tuple> results = queryFactory
-                .select(
-                        workMatch.status,
-                        workMatch.settlementAmount.sum(),
-                        workMatch.count()
-                )
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(caregiverCenter.center.centerId.eq(centerId))
-                .groupBy(workMatch.status)
-                .fetch();
-
-        BigDecimal totalAmount = BigDecimal.ZERO;
-        long completedCount = 0L;
-        long plannedCount = 0L;
-        long cancelledCount = 0L;
-
-        for (Tuple tuple : results) {
-            WorkStatus status = tuple.get(workMatch.status);
-            BigDecimal sumAmount = tuple.get(workMatch.settlementAmount.sum());
-            Long count = tuple.get(workMatch.count());
-
-            if (status == WorkStatus.COMPLETED) {
-                completedCount = count;
-                totalAmount = totalAmount.add(sumAmount != null ? sumAmount : BigDecimal.ZERO);
-            } else if (status == WorkStatus.PLANNED) {
-                plannedCount = count;
-            } else if (status == WorkStatus.CANCELLED) {
-                cancelledCount = count;
-            }
-        }
-
-        return new GetSettlementCenterSummaryResponse(
-                totalAmount,
-                completedCount,
-                plannedCount,
-                cancelledCount
-        );
-    }
-
-    // 요양보호사 개별 정산 금액, 건수 조회
-    public GetCaregiverSettlementSummaryResponse getCaregiverSettlementSummary(UUID caregiverId) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-
-        // 총 정산 금액 (isPaid = true만)
-        BigDecimal totalAmount = Optional.ofNullable(
-                queryFactory
-                        .select(workMatch.settlementAmount.sum())
-                        .from(workMatch)
-                        .where(
-                                workMatch.caregiver.caregiverId.eq(caregiverId),
-                                workMatch.status.eq(WorkStatus.COMPLETED),
-                                workMatch.isPaid.eq(true)
-                        )
-                        .fetchOne()
-        ).orElse(BigDecimal.ZERO);
-
-        // 상태별 카운트
-        long completedCount = Optional.ofNullable(
-                queryFactory
-                        .select(workMatch.count())
-                        .from(workMatch)
-                        .where(
-                                workMatch.caregiver.caregiverId.eq(caregiverId),
-                                workMatch.status.eq(WorkStatus.COMPLETED)
-                        )
-                        .fetchOne()
-        ).orElse(0L);
-
-        long plannedCount = Optional.ofNullable(
-                queryFactory
-                        .select(workMatch.count())
-                        .from(workMatch)
-                        .where(
-                                workMatch.caregiver.caregiverId.eq(caregiverId),
-                                workMatch.status.eq(WorkStatus.PLANNED)
-                        )
-                        .fetchOne()
-        ).orElse(0L);
-
-        long cancelledCount = Optional.ofNullable(
-                queryFactory
-                        .select(workMatch.count())
-                        .from(workMatch)
-                        .where(
-                                workMatch.caregiver.caregiverId.eq(caregiverId),
-                                workMatch.status.eq(WorkStatus.CANCELLED)
-                        )
-                        .fetchOne()
-        ).orElse(0L);
-
-        return new GetCaregiverSettlementSummaryResponse(
-                totalAmount,
-                completedCount,
-                plannedCount,
-                cancelledCount
-        );
-    }
-
-    // 센터에 등록된 요양보호사의 근무 상태, 월-연도별 내역 조회
-    public List<GetCaregiverWorkResponse> getCaregiverWorkListByCenter(
-            UUID centerId,
-            WorkStatus status,   // nullable
-            Integer year,        // nullable
-            Integer month        // nullable
-    ) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-        QUser user = QUser.user;
-
-        BooleanBuilder where = new BooleanBuilder();
-        where.and(caregiverCenter.center.centerId.eq(centerId));
-
-        // 상태 필터
-        if (status != null) {
-            where.and(workMatch.status.eq(status));
-        } else {
-            where.and(workMatch.status.eq(WorkStatus.COMPLETED));
-        }
-
-        // 연/월 필터
-        if (year != null && month != null) {
-            where.and(workMatch.workDate.year().eq(year)
-                    .and(workMatch.workDate.month().eq(month)));
-        }
-
-        return queryFactory
-                .select(Projections.constructor(
-                        GetCaregiverWorkResponse.class,
-                        user.name,
-                        workMatch.workDate,
-                        workMatch.workStartTime,
-                        workMatch.workEndTime,
-                        workMatch.settlementAmount,
-                        workMatch.status
-                ))
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiver.user, user)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(where)
-                .orderBy(
-                        workMatch.modifiedAt
-                                .coalesce(workMatch.createdAt)
-                                .desc()
-                )
-                .fetch();
-    }
-
-    // 개별 요양보호사의 근무 상태, 월-연도별 내역 조회
-    public List<GetCaregiverWorkResponse> getCaregiverWorkListByCaregiver(
-            UUID caregiverId,
-            WorkStatus status,   // nullable
-            Integer year,        // nullable
-            Integer month        // nullable
-    ) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-        QUser user = QUser.user;
-
-        BooleanBuilder where = new BooleanBuilder();
-        where.and(caregiver.caregiverId.eq(caregiverId));
-
-        // 상태 필터
-        if (status != null) {
-            where.and(workMatch.status.eq(status));
-        }
-
-        // 연/월 필터
-        if (year != null && month != null) {
-            where.and(workMatch.workDate.year().eq(year)
-                    .and(workMatch.workDate.month().eq(month)));
-        }
-
-        return queryFactory
-                .select(Projections.constructor(
-                        GetCaregiverWorkResponse.class,
-                        user.name,
-                        workMatch.workDate,
-                        workMatch.workStartTime,
-                        workMatch.workEndTime,
-                        workMatch.settlementAmount,
-                        workMatch.status
-                ))
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiver.user, user)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(where)
-                .orderBy(
-                        workMatch.modifiedAt
-                                .coalesce(workMatch.createdAt)
-                                .desc()
-                )
-                .fetch();
-    }
-
-    // 이번 달 총 정산내역 조회
-    public List<GetMonthlyPaymentResponse> getMonthlyPaidSettlements(UUID centerId, int monthsBack) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        LocalDate now = LocalDate.now();
-        LocalDate startMonth = now.minusMonths(monthsBack - 1).withDayOfMonth(1);
-
-        List<GetMonthlyPaymentResponse> rawResults = queryFactory
-                .select(Projections.constructor(
-                        GetMonthlyPaymentResponse.class,
-                        workMatch.workDate.year(),
-                        workMatch.workDate.month(),
-                        workMatch.settlementAmount.sum()
-                ))
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(workMatch.isPaid.eq(true)
-                        .and(workMatch.status.eq(WorkStatus.COMPLETED))
-                        .and(caregiverCenter.center.centerId.eq(centerId))
-                        .and(workMatch.workDate.goe(startMonth)))
-                .groupBy(workMatch.workDate.year(), workMatch.workDate.month())
-                .orderBy(workMatch.workDate.year().desc(), workMatch.workDate.month().desc())
-                .fetch();
-
-        // 누락된 월 채우기
-        Map<String, BigDecimal> map = rawResults.stream()
-                .collect(Collectors.toMap(
-                        r -> r.year() + "-" + r.month(),
-                        GetMonthlyPaymentResponse::totalAmount
-                ));
-
-        List<GetMonthlyPaymentResponse> filled = new ArrayList<>();
-        for (int i = 0; i < monthsBack; i++) {
-            LocalDate target = now.minusMonths(i);
-            String key = target.getYear() + "-" + target.getMonthValue();
-            filled.add(new GetMonthlyPaymentResponse(
-                    target.getYear(),
-                    target.getMonthValue(),
-                    map.getOrDefault(key, BigDecimal.ZERO)
-            ));
-        }
-
-        return filled;
-    }
-
-    // 일주일 간 미정산 내역 조회
-    public List<GetDailyUnsettledResponse> getDailyUnsettledCount(UUID centerId) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        LocalDate today = LocalDate.now();
-        LocalDate startDate = today.minusDays(6); // 최근 7일
-
-        // 실제 DB에서 미정산 내역 조회
-        List<GetDailyUnsettledResponse> rawResults = queryFactory
-                .select(Projections.constructor(
-                        GetDailyUnsettledResponse.class,
-                        workMatch.workDate,
-                        workMatch.count(),
-                        workMatch.settlementAmount.sum()
-                ))
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(workMatch.isPaid.eq(false)
-                        .and(workMatch.status.ne(WorkStatus.COMPLETED))
-                        .and(caregiverCenter.center.centerId.eq(centerId))
-                        .and(workMatch.workDate.between(startDate, today))
-                )
-                .groupBy(workMatch.workDate)
-                .orderBy(workMatch.workDate.asc())
-                .fetch();
-
-        // 누락된 날짜 채우기
-        Map<LocalDate, GetDailyUnsettledResponse> map = rawResults.stream()
-                .collect(Collectors.toMap(GetDailyUnsettledResponse::date, r -> r));
-
-        List<GetDailyUnsettledResponse> filled = new ArrayList<>();
-        for (int i = 0; i <= 6; i++) {
-            LocalDate date = startDate.plusDays(i);
-            filled.add(map.getOrDefault(date, new GetDailyUnsettledResponse(date, 0L, BigDecimal.ZERO)));
-        }
-
-        return filled;
-    }
-
-    public List<GetWorkMatchByDateResponse> findWorkMatchByDate(UUID centerId, LocalDate date) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        return queryFactory
-                .select(Projections.constructor(
-                        GetWorkMatchByDateResponse.class,
-                        workMatch.workMatchId,
-                        workMatch.workDate,
-                        workMatch.workStartTime,
-                        workMatch.workEndTime,
-                        caregiver.user.name
-                ))
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(workMatch.workDate.eq(date)
-                        .and(caregiverCenter.center.centerId.eq(centerId)))
-                .orderBy(workMatch.workStartTime.asc())
-                .fetch();
-    }
-
-    public List<GetWorkMatchByPaid> findWorkMatchByPaid(UUID centerId, Boolean isPaid) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QUser user = QUser.user;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        return queryFactory
-                .select(Projections.constructor(
-                        GetWorkMatchByPaid.class,
-                        workMatch.workMatchId,
-                        workMatch.workDate,
-                        caregiver.user.name
-                ))
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiver.user, user)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(workMatch.isPaid.eq(isPaid)
-                        .and(caregiverCenter.center.centerId.eq(centerId)))
-                .orderBy(
-                        workMatch.workDate.desc())
-                .fetch();
-    }
-
-
-    // 이번 달 누적 정산 금액 조회
-    public BigDecimal getTotalSettledAmountThisMonth(UUID centerId) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        LocalDate now = LocalDate.now();
-        LocalDate firstDay = now.withDayOfMonth(1);
-
-        return queryFactory
-                .select(workMatch.settlementAmount.sum())
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(workMatch.isPaid.eq(true)
-                        .and(workMatch.workDate.goe(firstDay))
-                        .and(caregiverCenter.center.centerId.eq(centerId)))
-                .fetchOne();
-    }
-
-    // 미정산 건수 조회
-    public Long countUnsettled(UUID centerId) {
-        QWorkMatch workMatch = QWorkMatch.workMatch;
-        QCaregiver caregiver = QCaregiver.caregiver;
-        QCaregiverCenter caregiverCenter = QCaregiverCenter.caregiverCenter;
-
-        return queryFactory
-                .select(workMatch.count())
-                .from(workMatch)
-                .join(workMatch.caregiver, caregiver)
-                .join(caregiverCenter).on(caregiverCenter.caregiver.eq(caregiver))
-                .where(workMatch.isPaid.eq(false)
-                        .and(caregiverCenter.center.centerId.eq(centerId)))
-                .fetchOne();
-    }
-}
