--- homecare/src/main/java/jaega/homecare/domain/workMatch/service/command/WorkMatchCommandService.java
@@ -1,73 +1,0 @@
-package jaega.homecare.domain.workMatch.service.command;
-
-import jaega.homecare.domain.workMatch.dto.req.CreateWorkMatchRequest;
-import jaega.homecare.domain.workMatch.entity.WorkMatch;
-import jaega.homecare.domain.workMatch.entity.WorkStatus;
-import jaega.homecare.domain.workMatch.mapper.WorkMatchMapper;
-import jaega.homecare.domain.workMatch.repository.WorkMatchRepository;
-import jaega.homecare.domain.caregiver.entity.Caregiver;
-import jaega.homecare.domain.caregiver.service.query.CaregiverQueryService;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.time.Duration;
-import java.time.LocalTime;
-import java.util.UUID;
-
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class WorkMatchCommandService {
-
-    private final WorkMatchRepository workMatchRepository;
-    private final CaregiverQueryService caregiverQueryService;
-    private final WorkMatchMapper workMatchMapper;
-
-
-    public void createWorkMatch(CreateWorkMatchRequest request){
-        Caregiver caregiver = caregiverQueryService.getCaregiver(request.caregiverId());
-
-        BigDecimal settlementAmount = calculateSettlementAmount(request.workStartTime(), request.workEndTime(), request.distanceLog());
-
-        WorkMatch workMatch = workMatchMapper.toEntity(request, caregiver, settlementAmount);
-        workMatch.initializeWorkMatch(UUID.randomUUID());
-        workMatchRepository.save(workMatch);
-    }
-
-    public BigDecimal calculateSettlementAmount(LocalTime start, LocalTime end, Double distanceKm) {
-        BigDecimal hourlyRate = new BigDecimal("15000");  // 시간당 단가
-        BigDecimal distanceRate = new BigDecimal("500");  // 거리당 단가
-
-        // 근무 시간 계산 (분 단위 -> 시간 단위로 변환)
-        long minutes = Duration.between(start, end).toMinutes();
-        BigDecimal hours = BigDecimal.valueOf(minutes).divide(BigDecimal.valueOf(60), 2, RoundingMode.HALF_UP);
-
-        // 거리
-        BigDecimal distance = BigDecimal.valueOf(distanceKm != null ? distanceKm : 0);
-
-        // 금액 계산
-        BigDecimal amount = hourlyRate.multiply(hours).add(distanceRate.multiply(distance));
-        return amount.setScale(0, RoundingMode.HALF_UP);  // 반올림, 소수점 없애기
-    }
-
-    /**
-     * 특정 WorkMatch의 상태 변경
-     */
-    public void changeWorkMatchStatus(UUID workMatchId, WorkStatus newStatus) {
-        WorkMatch workMatch = workMatchRepository.findByWorkMatchId(workMatchId)
-                .orElseThrow(() -> new IllegalArgumentException("WorkMatch not found: " + workMatchId));
-
-        workMatch.changeWorkStatus(newStatus);
-    }
-
-    public void changePaidStatus(UUID workMatchId) {
-        WorkMatch workMatch = workMatchRepository.findByWorkMatchId(workMatchId)
-                .orElseThrow(() -> new IllegalArgumentException("WorkLog not found: " + workMatchId));
-
-        workMatch.changePaidStatus();
-    }
-
-}
